extract endpoints& config
