+++ src_ts/elements/mixins/query-params-mixin.ts
@@ -1,4 +1,4 @@
-import {basePath} from '../app-shell/etools-app-config';
+import {basePath} from '../../config/config';
 
 export const parseQueries = () => {
   const queriesObj: any = {};
