+++ src_ts/elements/pages/action-points/action-point-details.ts
@@ -9,7 +9,7 @@ import '@unicef-polymer/etools-loading/etools-loading.js';
 import '@unicef-polymer/etools-date-time/datepicker-lite.js';
 import {LocalizationMixin} from '../../mixins/localization-mixin';
 import {InputAttrsMixin} from '../../mixins/input-attrs-mixin';
-import {getEndpoint} from '../../mixins/endpoint-mixin';
+import {getEndpoint} from '../../../endpoints/endpoint-mixin';
 import {DateMixin} from '../../mixins/date-mixin';
 import {getData} from '../../mixins/static-data-mixin';
 import {actionAllowed} from '../../mixins/permission-controller';
